--- src/main/java/com/example/videoworks/data/Appointment.java
@@ -1,10 +1,0 @@
-package com.example.videoworks.data;
-
-public class Appointment {
-
-    private Long id;
-    private String date;
-    private Double time;
-
-
-}
