--- src/main/java/com/example/videoworks/data/Category.java
@@ -1,32 +1,0 @@
-package com.example.videoworks.data;
-
-public class Category {
-
-    private Long id;
-    private String name;
-
-    public Category(Long id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public Category() {
-
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-}
